Apply sense changes to pre-1.5-13 master branch
Upgrading to a more recent version of upstream is necessary for R 4.x
support.

This commit copies the substantive changes made to the Sense version of
Cairo to the current tip of the master branch of upstream development
(which is post-1.5-13, but pre-1.5-13).

I have replaced the master branch of the SensePlatform/Cairo repo
with the master branch of the upstream repository. I renamed the
existing branches in the SensePlatform/Cairo repository:
{master,devel,1.5-6} -> sense-archive-{master,devel,1.5-6}.